refactor(models): promote key_registry and playbook_meta, delete models module
The `models` module is dissolved. With `playbook` inlined and
`inventory` co-located with its loader, `key_registry` and
`playbook_meta` are the only remaining occupants and both are
peer-level domain types alongside `config`, `user_config`, and
`hosts`. Promote them to top-level modules and drop `models.rs` and
the now-empty `src/models/` directory.

Closes #266
