+++ src/key_registry.rs
