--- src/models.rs
@@ -1,2 +1,0 @@
-pub mod key_registry;
-pub mod playbook_meta;
